% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,ANY-method}
\alias{extract,ANY-method}
\alias{extract.broom}
\alias{extract.ANY}
\alias{extract.ANY-method}
\title{\code{\link{extract}} method for \code{broom} objects}
\usage{
\S4method{extract}{ANY}(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{broom} objects created by the
\code{\link[broom]{broom}} function in the \pkg{broom} package.
}
