% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,apollo-method}
\alias{extract,apollo-method}
\alias{extract.apollo}
\title{\code{\link{extract}} method for \code{apollo} objects}
\usage{
\S4method{extract}{apollo}(model, wtpest = NULL, se = "rob", ...)
}
\arguments{
\item{model}{An object of class \code{apollo}.}

\item{wtpest}{Optional data.frame of willingness-to-pay estimates (and s.e.'s).}

\item{se}{Which standard errors to use: \code{"rob"}, \code{"normal"}, or \code{"bs"}.}

\item{...}{Currently ignored.}
}
\description{
\code{\link{extract}} method for \code{apollo} objects created by
\code{apollo::apollo_estimate()}.
}
\author{
Julian Sagebiel <julian.sagebiel@idiv.de>
}
