% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,Arima-method}
\alias{extract,Arima-method}
\alias{extract.Arima}
\alias{extract,ARIMA-method}
\alias{extract.ARIMA}
\title{\code{\link{extract}} method for \code{Arima} objects}
\usage{
\S4method{extract}{Arima}(
  model,
  include.pvalues = TRUE,
  include.aic = TRUE,
  include.loglik = TRUE,
  ...
)

\S4method{extract}{ARIMA}(
  model,
  include.pvalues = FALSE,
  include.aic = TRUE,
  include.aicc = TRUE,
  include.bic = TRUE,
  include.loglik = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.pvalues}{Report p-values?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}

\item{include.aicc}{Report AICC in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}
}
\description{
\code{\link{extract}} method for \code{Arima} objects created by the
\code{\link[stats]{arima}} function in the \pkg{stats} package.

\code{\link{extract}} method for \code{ARIMA} objects created by the
\code{\link[forecast]{Arima}} function in the \pkg{forecast} package.
}
