% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,bergm-method}
\alias{extract,bergm-method}
\alias{extract.bergm}
\title{\code{\link{extract}} method for \code{bergm} objects}
\usage{
\S4method{extract}{bergm}(model, posterior.median = FALSE, level = 0.95, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{posterior.median}{Report the posterior median instead of the default
posterior mean as coefficients?}

\item{level}{Confidence level, i.e., the proportion of the posterior
distribution to be included in the credible interval.}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{bergm} objects created by the
\code{\link[Bergm]{bergm}} function in the \pkg{Bergm} package.
}
