% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,betamfx-method}
\alias{extract,betamfx-method}
\alias{extract.betamfx}
\title{\code{\link{extract}} method for \code{betamfx} objects}
\usage{
\S4method{extract}{betamfx}(
  model,
  include.pseudors = TRUE,
  include.loglik = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.pseudors}{Report pseudo R^2 in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{betamfx} objects created by the
\code{\link[mfx]{betamfx}} function in the \pkg{mfx} package.
}
