% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,bife-method}
\alias{extract,bife-method}
\alias{extract.bife}
\title{\code{\link{extract}} method for \code{bife} objects}
\usage{
\S4method{extract}{bife}(
  model,
  include.loglik = TRUE,
  include.deviance = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.deviance}{Report the residual deviance?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{bife} objects created by the
\code{\link[bife]{bife}} function in the \pkg{bife} package.
}
\author{
Philip Leifeld, Christoph Riedl, Claudia Zucca
}
