% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,biglm-method}
\alias{extract,biglm-method}
\alias{extract.biglm}
\title{\code{\link{extract}} method for \code{biglm} objects}
\usage{
\S4method{extract}{biglm}(model, include.nobs = TRUE, include.aic = TRUE, use.ci = FALSE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{use.ci}{Report confidence intervals in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{biglm} objects created by the
\code{\link[biglm]{biglm}} function in the \pkg{biglm} package.
}
\author{
Claudia Zucca, Philip Leifeld
}
