% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,btergm-method}
\alias{extract,btergm-method}
\alias{extract.btergm}
\title{\code{\link{extract}} method for \code{btergm} objects}
\usage{
\S4method{extract}{btergm}(model, level = 0.95, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{level}{Significance or confidence level (\code{1 - alpha}) for
computing confidence intervals.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{btergm} objects created by the
\code{\link[btergm]{btergm}} function in the \pkg{btergm} package.
}
