% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,coeftest-method}
\alias{extract,coeftest-method}
\alias{extract.coeftest}
\title{\code{\link{extract}} method for \code{coeftest} objects}
\usage{
\S4method{extract}{coeftest}(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{coeftest} objects created by the
\code{\link[lmtest]{coeftest}} function in the \pkg{lmtest} package.
}
