% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,ergm-method}
\alias{extract,ergm-method}
\alias{extract.ergm}
\title{\code{\link{extract}} method for \code{ergm} objects}
\usage{
\S4method{extract}{ergm}(
  model,
  include.aic = TRUE,
  include.bic = TRUE,
  include.loglik = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{ergm} objects created by the
\code{\link[ergm]{ergm}} function in the \pkg{ergm} package.
}
