% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,ergmm-method}
\alias{extract,ergmm-method}
\alias{extract.ergmm}
\title{\code{\link{extract}} method for \code{ergmm} objects}
\usage{
\S4method{extract}{ergmm}(model, include.bic = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{ergmm} objects created by the
\code{\link[latentnet]{ergmm}} function in the \pkg{latentnet}
package.
}
