% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,fc_model-method}
\alias{extract,fc_model-method}
\alias{extract.fc_model}
\title{\code{\link{extract}} method for \code{fc_model} objects}
\usage{
\S4method{extract}{fc_model}(
  model,
  include.pvalues = TRUE,
  include.aic = TRUE,
  include.aicc = TRUE,
  include.bic = TRUE,
  include.loglik = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.pvalues}{Report p-values?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.aicc}{Report AICC in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{fc_model} objects created by the
\code{\link[forecast]{Arima}} function in the \pkg{forecast} package.
}
