% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,feglm-method}
\alias{extract,feglm-method}
\alias{extract.feglm}
\title{\code{\link{extract}} method for \code{feglm} objects}
\usage{
\S4method{extract}{feglm}(
  model,
  include.deviance = TRUE,
  include.nobs = TRUE,
  include.groups = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.deviance}{Report the deviance?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.groups}{Report the number of groups?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{feglm} objects created by the
\code{\link[alpaca]{feglm}} function in the \pkg{alpaca} package.
}
