% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,fGARCH-method}
\alias{extract,fGARCH-method}
\alias{extract.fGARCH}
\title{\code{\link{extract}} method for \code{fGARCH} objects}
\usage{
\S4method{extract}{fGARCH}(
  model,
  include.nobs = TRUE,
  include.aic = TRUE,
  include.loglik = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{fGARCH} objects created by the
\code{\link[fGarch]{garchFit}} function in the \pkg{fGarch} package.
}
