% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,fixest-method}
\alias{extract,fixest-method}
\alias{extract.fixest}
\title{\code{\link{extract}} method for \code{fixest} objects}
\usage{
\S4method{extract}{fixest}(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{\link[fixest]{coeftable}} method for the object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.proj.stats}{Include statistics for projected model in the GOF block?}

\item{include.deviance}{Report the deviance in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.groups}{Report the number of groups?}

\item{include.pseudors}{Report Pseudo-R^2 in the GOF block?}
}
\description{
\code{\link{extract}} method for model objects created by the
model fitting functions (e.g. \code{\link[fixes]{feols}},
\code{\link[fixes]{feglm}}) in the \pkg{fixest} package.
}
\details{
Note that appropriate arguments depend on the model fitting method used.
For example, \code{include.pseudors} does not make sense for models fit
with \code{\link[fixest]{feols}}.
}
\author{
Christopher Poliquin
}
