% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,forecast-method}
\alias{extract,forecast-method}
\alias{extract.forecast}
\title{\code{\link{extract}} method for \code{forecast} objects}
\usage{
\S4method{extract}{forecast}(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{forecast} objects created by the
\code{\link[forecast]{forecast}} and \code{\link[forecast:ses]{holt}} functions
in the \pkg{forecast} package.
}
