% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,gamlss-method}
\alias{extract,gamlss-method}
\alias{extract.gamlss}
\title{\code{\link{extract}} method for \code{gamlss} objects}
\usage{
\S4method{extract}{gamlss}(model, robust = FALSE, include.nobs = TRUE,
  include.nagelkerke = TRUE, include.gaic = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{robust}{If TRUE computes robust standard errors in the variance covariance matrix.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.nagelkerke}{Report Nagelkerke R^2 in the GOF block?}

\item{include.gaic}{Report Generalized Akaike's Information Criterion (AIC) in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{gamlss} objects. These objects are
created by the \code{\link[gamlss]{gamlss}} function in the \pkg{gamlss} package.
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,gel-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,negbin-method}},
  \code{\link{extract,speedglm-method}},
  \code{\link{extract,speedlm-method}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.lme4}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
