% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,gamlssZadj-method}
\alias{extract,gamlssZadj-method}
\alias{extract.gamlssZadj}
\title{\code{\link{extract}} method for \code{gamlssZadj} objects}
\usage{
\S4method{extract}{gamlssZadj}(
  model,
  type = c("qr", "vcov"),
  robust = FALSE,
  include.nobs = TRUE,
  include.gaic = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{type}{The type.}

\item{robust}{If TRUE computes robust standard errors in the
variance-covariance matrix.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.gaic}{Report Generalized Akaike's Information Criterion (AIC)
in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{vcov} method for the object.}
}
\description{
\code{\link{extract}} method for \code{gamlssZadj} objects created by the
\code{\link[gamlss.inf]{gamlssZadj}} function in the \pkg{gamlss.inf}
package.
}
