% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,geeglm-method}
\alias{extract,geeglm-method}
\alias{extract.geeglm}
\title{\code{\link{extract}} method for \code{geeglm} objects}
\usage{
\S4method{extract}{geeglm}(
  model,
  include.scale = TRUE,
  include.correlation = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.scale}{Report the dispersion or scale parameter?}

\item{include.correlation}{Report the correlation parameter alpha and its
standard error?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{geeglm} objects created by the
\code{\link[geepack]{geeglm}} function in the \pkg{geepack} package.
}
