% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,glm-method}
\alias{extract,glm-method}
\alias{extract.glm}
\title{\code{\link{extract}} method for \code{glm} objects}
\usage{
\S4method{extract}{glm}(model, include.aic = TRUE, include.bic = TRUE,
  include.loglik = TRUE, include.deviance = TRUE,
  include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.deviance}{Report the deviance in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function
for \code{glm} objects.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{glm} objects. These objects are
created by the \code{\link[stats]{glm}} function in the \pkg{stats} package.
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,H2OBinomialModel-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,gnm-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gmm}}, \code{\link{extract.lme4}},
  \code{\link{extract.mfx}}, \code{\link{extract.mgcv}},
  \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
