% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,glmrob-method}
\alias{extract,glmrob-method}
\alias{extract.glmrob}
\title{\code{\link{extract}} method for \code{glmrob} objects}
\usage{
\S4method{extract}{glmrob}(model, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{glmrob} objects created by the
\code{\link[robustbase]{glmrob}} function in the \pkg{robustbase} package.
}
