% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,gmm-method}
\alias{extract,gmm-method}
\alias{extract.gmm}
\title{\code{\link{extract}} method for \code{gmm} objects}
\usage{
\S4method{extract}{gmm}(
  model,
  include.obj.fcn = TRUE,
  include.overidentification = FALSE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.obj.fcn}{Report the value of the objective function
(= criterion function)? More precisely, this returns
\code{E(g)var(g)^{-1}E(g)}.}

\item{include.overidentification}{Report the J-test for overidentification?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{gmm} objects created by the
\code{\link[gmm]{gmm}} function in the \pkg{gmm} package.
}
