% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,lm-method}
\alias{extract,lm-method}
\alias{extract.lm}
\title{\code{\link{extract}} method for \code{lm} objects}
\usage{
\S4method{extract}{lm}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE,
  include.fstatistic = FALSE, include.rmse = FALSE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.fstatistic}{Report the F-statistic in the GOF block?}

\item{include.rmse}{Report the root mean square error (RMSE; = residual
standard deviation) in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function
for \code{lm} objects.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{lm} objects. These objects are
created by the \code{\link[stats]{lm}} function in the \pkg{stats} package.
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,ANY-method}},
  \code{\link{extract,ARIMA-method}},
  \code{\link{extract,Arima-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betamfx-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,btergm-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,clm-method}},
  \code{\link{extract,clmm-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,coxph-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,ets-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lme4-method}},
  \code{\link{extract,negbin-method}},
  \code{\link{extract,sclm-method}},
  \code{\link{extract,speedglm-method}},
  \code{\link{extract,speedlm-method}}
}
\author{
Philip Leifeld
}
\concept{extract}
