% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,lrm-method}
\alias{extract,lrm-method}
\alias{extract.lrm}
\title{\code{\link{extract}} method for \code{lrm} objects}
\usage{
\S4method{extract}{lrm}(
  model,
  include.pseudors = TRUE,
  include.lr = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.pseudors}{Report pseudo R^2 in the GOF block?}

\item{include.lr}{Report likelihood ratio test?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{lrm} objects created by the
\code{\link[rms]{lrm}} function in the \pkg{rms} package.
}
\author{
Fabrice Le Lec
}
