% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,maBina-method}
\alias{extract,maBina-method}
\alias{extract.maBina}
\title{\code{\link{extract}} method for \code{maBina} objects}
\usage{
\S4method{extract}{maBina}(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines.}
}
\description{
\code{\link{extract}} method for \code{maBina} objects created by the
\code{\link[erer]{maBina}} function in the \pkg{erer} package.
}
