% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,maxLik-method}
\alias{extract,maxLik-method}
\alias{extract.maxLik}
\title{\code{\link{extract}} method for \code{maxLik} objects}
\usage{
\S4method{extract}{maxLik}(model, include.logLik = TRUE, include.aic = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.aic}{Report the AIC in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}

\item{include.loglik}{Report the log likelihood in the GOF block?}
}
\description{
\code{\link{extract}} method for \code{maxLik} objects created by the
\code{\link[maxLik]{maxLik}} function in the \pkg{maxLik} package.
}
