% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,mhurdle-method}
\alias{extract,mhurdle-method}
\alias{extract.mhurdle}
\title{\code{\link{extract}} method for \code{mhurdle} objects}
\usage{
\S4method{extract}{mhurdle}(model, include.nobs = TRUE, include.loglik = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{mhurdle} objects created by the
\code{\link[mhurdle]{mhurdle}} function in the \pkg{mhurdle} package.
}
