% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,mlogit-method}
\alias{extract,mlogit-method}
\alias{extract.mlogit}
\title{\code{\link{extract}} method for \code{mlogit} objects}
\usage{
\S4method{extract}{mlogit}(
  model,
  include.aic = TRUE,
  include.loglik = TRUE,
  include.nobs = TRUE,
  include.groups = TRUE,
  include.order = FALSE,
  include.iterations = FALSE,
  beside = FALSE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.groups}{Report the number of groups?}

\item{include.order}{Report coefficient names in alphabetical order?}

\item{include.iterations}{Report the number of iterations?}

\item{beside}{Arrange the model terms below each other or beside each other?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{mlogit} objects created by the
\code{\link[mlogit]{mlogit}} function in the \pkg{mlogit} package.
}
