% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,negbinirr-method}
\alias{extract,negbinirr-method}
\alias{extract.negbinirr}
\title{\code{\link{extract}} method for \code{negbinirr} objects}
\usage{
\S4method{extract}{negbinirr}(
  model,
  include.nobs = TRUE,
  include.loglik = TRUE,
  include.deviance = TRUE,
  include.aic = TRUE,
  include.bic = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.deviance}{Report the deviance?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{negbinirr} objects created by the
\code{\link[mfx]{negbinirr}} function in the \pkg{mfx} package.
}
