% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,netlogit-method}
\alias{extract,netlogit-method}
\alias{extract.netlogit}
\title{\code{\link{extract}} method for \code{netlogit} objects}
\usage{
\S4method{extract}{netlogit}(
  model,
  include.aic = TRUE,
  include.bic = TRUE,
  include.deviance = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{include.deviance}{Report the deviance?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{netlogit} objects created by the
\code{netlogit} function in the \pkg{sna} package.
}
