% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,nlmerMod-method}
\alias{extract,nlmerMod-method}
\alias{extract.nlmerMod}
\title{\code{\link{extract}} method for \code{nlmerMod} objects}
\usage{
\S4method{extract}{nlmerMod}(
  model,
  method = c("naive", "profile", "boot", "Wald"),
  level = 0.95,
  nsim = 1000,
  include.aic = TRUE,
  include.bic = TRUE,
  include.dic = FALSE,
  include.deviance = FALSE,
  include.loglik = TRUE,
  include.nobs = TRUE,
  include.groups = TRUE,
  include.variance = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{method}{The method used to compute confidence intervals or p-values.
The default value \code{"naive"} computes naive p-values while the other
methods compute confidence intervals using the \code{confint} function. See
\code{\link[lme4]{confint.merMod}}.}

\item{level}{Significance or confidence level (\code{1 - alpha}) for
computing confidence intervals.}

\item{nsim}{The MCMC sample size or number of bootstrapping replications on
the basis of which confidence intervals are computed (only if the
\code{method} argument does not specify \code{"naive"}, which is the
default behavior). Note: large values may take considerable computing time.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{include.dic}{Report the deviance information criterion (DIC)?}

\item{include.deviance}{Report the deviance?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.groups}{Report the number of groups?}

\item{include.variance}{Report group variances?}

\item{...}{Arguments to be passed to the \code{\link[lme4]{fixef}} function
in the \pkg{lme4} package.}
}
\description{
\code{\link{extract}} method for \code{nlmerMod} objects created by the
\code{\link[lme4]{nlmer}} function in the \pkg{lme4} package.
}
\seealso{
Other extract: 
\code{\link{extract,ANY-method}},
\code{\link{extract,ARIMA-method}},
\code{\link{extract,Arima-method}},
\code{\link{extract,H2OBinomialModel-method}},
\code{\link{extract,Zelig-method}},
\code{\link{extract,aftreg-method}},
\code{\link{extract,averaging-method}},
\code{\link{extract,bam-method}},
\code{\link{extract,betamfx-method}},
\code{\link{extract,betaor-method}},
\code{\link{extract,betareg-method}},
\code{\link{extract,bife-method}},
\code{\link{extract,biglm-method}},
\code{\link{extract,brglm-method}},
\code{\link{extract,brmsfit-method}},
\code{\link{extract,btergm-method}},
\code{\link{extract,censReg-method}},
\code{\link{extract,clm-method}},
\code{\link{extract,clmm-method}},
\code{\link{extract,clogit-method}},
\code{\link{extract,coeftest-method}},
\code{\link{extract,coxph-method}},
\code{\link{extract,coxph.penal-method}},
\code{\link{extract,coxreg-method}},
\code{\link{extract,dynlm-method}},
\code{\link{extract,ergm-method}},
\code{\link{extract,ergmm-method}},
\code{\link{extract,ets-method}},
\code{\link{extract,fGARCH-method}},
\code{\link{extract,feglm-method}},
\code{\link{extract,feis-method}},
\code{\link{extract,felm-method}},
\code{\link{extract,forecast-method}},
\code{\link{extract,gam-method}},
\code{\link{extract,gamlss-method}},
\code{\link{extract,gee-method}},
\code{\link{extract,geeglm-method}},
\code{\link{extract,gel-method}},
\code{\link{extract,glm-method}},
\code{\link{extract,glmerMod-method}},
\code{\link{extract,glmmPQL-method}},
\code{\link{extract,glmmadmb-method}},
\code{\link{extract,glmrob-method}},
\code{\link{extract,gls-method}},
\code{\link{extract,gmm-method}},
\code{\link{extract,gnls-method}},
\code{\link{extract,gnm-method}},
\code{\link{extract,hurdle-method}},
\code{\link{extract,lm-method}},
\code{\link{extract,lmRob-method}},
\code{\link{extract,lme-method}},
\code{\link{extract,lme4-method}},
\code{\link{extract,lmerMod-method}},
\code{\link{extract,lmrob-method}},
\code{\link{extract,lnam-method}},
\code{\link{extract,logitmfx-method}},
\code{\link{extract,logitor-method}},
\code{\link{extract,lqmm-method}},
\code{\link{extract,lrm-method}},
\code{\link{extract,maBina-method}},
\code{\link{extract,merMod-method}},
\code{\link{extract,mhurdle-method}},
\code{\link{extract,mlogit-method}},
\code{\link{extract,mnlogit-method}},
\code{\link{extract,model.selection-method}},
\code{\link{extract,mtergm-method}},
\code{\link{extract,multinom-method}},
\code{\link{extract,negbin-method}},
\code{\link{extract,negbinirr-method}},
\code{\link{extract,negbinmfx-method}},
\code{\link{extract,netlogit-method}},
\code{\link{extract,nlme-method}},
\code{\link{extract,oglmx-method}},
\code{\link{extract,ols-method}},
\code{\link{extract,panelAR-method}},
\code{\link{extract,pglm-method}},
\code{\link{extract,pgmm-method}},
\code{\link{extract,phreg-method}},
\code{\link{extract,plm-method}},
\code{\link{extract,pmg-method}},
\code{\link{extract,poissonirr-method}},
\code{\link{extract,poissonmfx-method}},
\code{\link{extract,polr-method}},
\code{\link{extract,probitmfx-method}},
\code{\link{extract,rem.dyad-method}},
\code{\link{extract,rlm-method}},
\code{\link{extract,rq-method}},
\code{\link{extract,sarlm-method}},
\code{\link{extract,sclm-method}},
\code{\link{extract,selection-method}},
\code{\link{extract,sienaFit-method}},
\code{\link{extract,simex-method}},
\code{\link{extract,speedglm-method}},
\code{\link{extract,speedlm-method}},
\code{\link{extract,stergm-method}},
\code{\link{extract,summary.lm-method}},
\code{\link{extract,survreg-method}},
\code{\link{extract,survreg.penal-method}},
\code{\link{extract,svyglm-method}},
\code{\link{extract,systemfit-method}},
\code{\link{extract,texreg-method}},
\code{\link{extract,tobit-method}},
\code{\link{extract,vglm-method}},
\code{\link{extract,weibreg-method}},
\code{\link{extract,wls-method}},
\code{\link{extract,zelig-method}},
\code{\link{extract,zeroinfl-method}},
\code{\link{extract}()}
}
\concept{extract}
