% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,oglmx-method}
\alias{extract,oglmx-method}
\alias{extract.oglmx}
\title{\code{\link{extract}} method for \code{oglmx} objects}
\usage{
\S4method{extract}{oglmx}(
  model,
  include.aic = TRUE,
  include.iterations = TRUE,
  include.loglik = TRUE,
  include.nobs = TRUE,
  include.rsquared = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.iterations}{Report the number of iterations?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{oglmx} objects created by the
\code{\link[oglmx]{oglmx}} function in the \pkg{oglmx} package.
}
