% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,ols-method}
\alias{extract,ols-method}
\alias{extract.ols}
\title{\code{\link{extract}} method for \code{ols} objects}
\usage{
\S4method{extract}{ols}(
  model,
  include.nobs = TRUE,
  include.rsquared = TRUE,
  include.adjrs = TRUE,
  include.fstatistic = FALSE,
  include.lr = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.fstatistic}{Report the F-statistic in the GOF block?}

\item{include.lr}{Report likelihood ratio test?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{ols} objects created by the
\code{\link[rms]{ols}} function in the \pkg{rms} package.
}
