% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,panelAR-method}
\alias{extract,panelAR-method}
\alias{extract.panelAR}
\title{\code{\link{extract}} method for \code{panelAR} objects}
\usage{
\S4method{extract}{panelAR}(
  model,
  include.rsquared = TRUE,
  include.nobs = TRUE,
  include.groups = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.groups}{Report the number of groups?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{panelAR} objects created by the
\code{\link[panelAR]{panelAR}} function in the \pkg{panelAR} package.
}
