% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,pcce-method}
\alias{extract,pcce-method}
\alias{extract.pcce}
\title{\code{\link{extract}} method for \code{pcce} objects}
\usage{
\S4method{extract}{pcce}(
  model,
  include.r.squared = TRUE,
  include.sumsquares = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.r.squared}{Report the HPY R-squared statistic in the GOF block?}

\item{include.sumsquares}{Report the total and residual sum of squares in the
GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{pcce} objects created by the
\code{\link[plm]{pcce}} function in the \pkg{plm} package.
}
