% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,pgmm-method}
\alias{extract,pgmm-method}
\alias{extract.pgmm}
\title{\code{\link{extract}} method for \code{pgmm} objects}
\usage{
\S4method{extract}{pgmm}(
  model,
  include.nobs = TRUE,
  include.sargan = TRUE,
  include.wald = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.sargan}{Report the Sargan test?}

\item{include.wald}{Report the Wald statistic?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{pgmm} objects created by the
\code{\link[plm]{pgmm}} function in the \pkg{plm} package.
}
