% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,plm-method}
\alias{extract,plm-method}
\alias{extract.plm}
\title{\code{\link{extract}} method for \code{plm} objects}
\usage{
\S4method{extract}{plm}(
  model,
  include.rsquared = TRUE,
  include.adjrs = TRUE,
  include.nobs = TRUE,
  include.variance = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.variance}{Report group variances?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{plm} objects created by the
\code{\link[plm]{plm}} function in the \pkg{plm} package.
}
