% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,prais-method}
\alias{extract,prais-method}
\alias{extract.prais}
\title{\code{\link{extract}} method for \code{prais} objects}
\usage{
\S4method{extract}{prais}(
  model,
  include.rsquared = TRUE,
  include.adjrs = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{prais} objects created by the
\code{\link[prais]{prais_winsten}} function in the \pkg{prais} package.
}
