% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,rem.dyad-method}
\alias{extract,rem.dyad-method}
\alias{extract.rem.dyad}
\title{\code{\link{extract}} method for \code{rem.dyad} objects}
\usage{
\S4method{extract}{rem.dyad}(
  model,
  include.nvertices = TRUE,
  include.events = TRUE,
  include.aic = TRUE,
  include.aicc = TRUE,
  include.bic = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nvertices}{Report the number of vertices in a STERGM?}

\item{include.events}{Report the number of events in the GOF block?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.aicc}{Report AICC in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{rem.dyad} objects created by the
\code{\link[relevent]{rem.dyad}} function in the \pkg{relevent} package.
}
