% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,rlm-method}
\alias{extract,rlm-method}
\alias{extract.rlm}
\title{\code{\link{extract}} method for \code{rlm} objects}
\usage{
\S4method{extract}{rlm}(model, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{rlm} objects created by the
\code{\link[MASS]{rlm}} function in the \pkg{MASS} package.
}
