% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,rq-method}
\alias{extract,rq-method}
\alias{extract.rq}
\title{\code{\link{extract}} method for \code{rq} objects}
\usage{
\S4method{extract}{rq}(model, include.nobs = TRUE, include.percentile = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.percentile}{Report the percentile (tau)?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{rq} objects created by the
\code{rq} function in the \pkg{quantreg} package.
}
