% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,sienaFit-method}
\alias{extract,sienaFit-method}
\alias{extract.sienaFit}
\title{\code{\link{extract}} method for \code{sienaFit} objects}
\usage{
\S4method{extract}{sienaFit}(model, include.iterations = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.iterations}{Report the number of iterations?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{sienaFit} objects created by the
\code{siena07} function in the \pkg{RSiena} package.
}
