% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,simex-method}
\alias{extract,simex-method}
\alias{extract.simex}
\title{\code{\link{extract}} method for \code{simex} objects}
\usage{
\S4method{extract}{simex}(model, jackknife = TRUE, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{jackknife}{Use Jackknife variance instead of asymptotic variance?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{simex} objects created by the
\code{\link[simex]{simex}} function in the \pkg{simex} package.
}
