% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,svyglm-method}
\alias{extract,svyglm-method}
\alias{extract.svyglm}
\title{\code{\link{extract}} method for \code{svyglm} objects}
\usage{
\S4method{extract}{svyglm}(
  model,
  include.aic = FALSE,
  include.bic = FALSE,
  include.loglik = FALSE,
  include.deviance = TRUE,
  include.dispersion = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF
block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.deviance}{Report the deviance?}

\item{include.dispersion}{Report the dispersion parameter?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{svyglm} objects created by the
\code{svyglm} function in the \pkg{survey} package.
}
