% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,systemfit-method}
\alias{extract,systemfit-method}
\alias{extract.systemfit}
\title{\code{\link{extract}} method for \code{systemfit} objects}
\usage{
\S4method{extract}{systemfit}(
  model,
  include.rsquared = TRUE,
  include.adjrs = TRUE,
  include.nobs = TRUE,
  beside = FALSE,
  include.suffix = FALSE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{beside}{Arrange the model terms below each other or beside each other,
in separate columns?}

\item{include.suffix}{Report the name of the current model in parentheses
after each model term (instead of before the model term)?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{systemfit} objects created by the
\code{\link[systemfit]{systemfit}} function in the \pkg{systemfit} package.
}
