% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,texreg-method}
\alias{extract,texreg-method}
\alias{extract.texreg}
\title{\code{\link{extract}} method for \code{texreg} objects}
\usage{
\S4method{extract}{texreg}(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{texreg} objects created by the
\code{\link[texreg]{extract}} function in the \pkg{texreg} package.
}
