% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,vglm-method}
\alias{extract,vglm-method}
\alias{extract.vglm}
\title{\code{\link{extract}} method for \code{vglm} objects}
\usage{
\S4method{extract}{vglm}(
  model,
  include.loglik = TRUE,
  include.df = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.df}{Report the degrees of freedom?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{vglm} objects created by the
\code{\link[VGAM]{vglm}} function in the \pkg{VGAM} package.
}
\author{
Christoph Riedl <c.riedl@neu.edu>
}
