% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,weibreg-method}
\alias{extract,weibreg-method}
\alias{extract.weibreg}
\title{\code{\link{extract}} method for \code{weibreg} objects}
\usage{
\S4method{extract}{weibreg}(
  model,
  include.aic = TRUE,
  include.loglik = TRUE,
  include.lr = TRUE,
  include.nobs = TRUE,
  include.events = TRUE,
  include.trisk = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF
block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.lr}{Report likelihood ratio test?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.events}{Report the number of events in the GOF block?}

\item{include.trisk}{Report the total time at risk (in event-history models)?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{weibreg} objects created by the
\code{\link[eha]{weibreg}} function in the \pkg{eha} package.
}
