% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.broom}
\alias{extract.broom}
\alias{extract.broom-methods}
\alias{extract,ANY-method}
\title{\code{\link{extract}} methods for models in the \pkg{broom} package}
\usage{
\S4method{extract}{ANY}(model, ...)
}
\arguments{
\item{model}{Can be any statistical model that is not covered by the
\code{\link{extract}} methods in \pkg{texreg} but is covered by the
\pkg{broom} package.}

\item{...}{Currently not in use.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for \code{broom} objects in the \pkg{broom}
package, in particular:
\itemize{
  \item These objects are created by the \code{\link[broom]{broom}} function 
  in the \pkg{broom} package.
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,coeftest-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.MuMIn}},
  \code{\link{extract.RSiena}}, \code{\link{extract.VGAM}},
  \code{\link{extract.Zelig}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.betareg}},
  \code{\link{extract.bife}}, \code{\link{extract.biglm}},
  \code{\link{extract.brglm}}, \code{\link{extract.brms}},
  \code{\link{extract.btergm}},
  \code{\link{extract.censReg}},
  \code{\link{extract.dynlm}}, \code{\link{extract.eha}},
  \code{\link{extract.erer}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.lqmm}},
  \code{\link{extract.metaSEM}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.mhurdle}},
  \code{\link{extract.mlogit}},
  \code{\link{extract.mnlogit}},
  \code{\link{extract.nlme}}, \code{\link{extract.nnet}},
  \code{\link{extract.oglmx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.panelAR}},
  \code{\link{extract.pglm}}, \code{\link{extract.plm}},
  \code{\link{extract.pscl}},
  \code{\link{extract.quantreg}},
  \code{\link{extract.relevent}},
  \code{\link{extract.rms}},
  \code{\link{extract.robustbase}},
  \code{\link{extract.robust}},
  \code{\link{extract.sampleSelection}},
  \code{\link{extract.simex}}, \code{\link{extract.sna}},
  \code{\link{extract.spatialreg}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survey}},
  \code{\link{extract.survival}},
  \code{\link{extract.systemfit}},
  \code{\link{extract.tergm}}, \code{\link{extract.texreg}}
}
\author{
Philip Leifeld
}
\concept{extract}
