% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.eha}
\alias{extract.eha}
\alias{extract.eha-methods}
\alias{extract,weibreg-method}
\alias{extract.weibreg}
\alias{extract,phreg-method}
\alias{extract.phreg}
\alias{extract,aftreg-method}
\alias{extract.aftreg}
\alias{extract,coxreg-method}
\alias{extract.coxreg}
\title{\code{\link{extract}} methods for models in the \pkg{eha} package}
\usage{
\S4method{extract}{weibreg}(model, include.aic = TRUE,
  include.loglik = TRUE, include.lr = TRUE, include.nobs = TRUE,
  include.events = TRUE, include.trisk = TRUE, ...)

\S4method{extract}{phreg}(model, include.aic = TRUE,
  include.loglik = TRUE, include.lr = TRUE, include.nobs = TRUE,
  include.events = TRUE, include.trisk = TRUE, ...)

\S4method{extract}{aftreg}(model, include.aic = TRUE,
  include.loglik = TRUE, include.lr = TRUE, include.nobs = TRUE,
  include.events = TRUE, include.trisk = TRUE, ...)

\S4method{extract}{coxreg}(model, include.aic = TRUE,
  include.loglik = TRUE, include.lr = TRUE, include.nobs = TRUE,
  include.events = TRUE, include.trisk = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.lr}{Report the model likelihood ration (L.R.) in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.events}{Report the number of events in the GOF block?}

\item{include.trisk}{Report the total time at risk in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{eha}
package, in particular:
\itemize{
\item \code{aftreg} objects created by the \code{\link[eha]{aftreg}}
    function in the \pkg{eha} package
  \item \code{coxreg} objects created by the \code{\link[eha]{coxreg}}
    function in the \pkg{eha} package
  \item \code{phreg} objects created by the \code{\link[eha]{phreg}}
    function in the \pkg{eha} package
  \item \code{weibreg} objects created by the \code{\link[eha]{weibreg}}
    function in the \pkg{eha} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,coeftest-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.MuMIn}},
  \code{\link{extract.RSiena}}, \code{\link{extract.VGAM}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.betareg}},
  \code{\link{extract.bife}}, \code{\link{extract.biglm}},
  \code{\link{extract.brglm}}, \code{\link{extract.brms}},
  \code{\link{extract.broom}},
  \code{\link{extract.btergm}},
  \code{\link{extract.censReg}},
  \code{\link{extract.dynlm}}, \code{\link{extract.erer}},
  \code{\link{extract.ergm}}, \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.lqmm}},
  \code{\link{extract.metaSEM}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.mhurdle}},
  \code{\link{extract.mlogit}},
  \code{\link{extract.mnlogit}},
  \code{\link{extract.nlme}}, \code{\link{extract.nnet}},
  \code{\link{extract.oglmx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.panelAR}},
  \code{\link{extract.pglm}}, \code{\link{extract.plm}},
  \code{\link{extract.quantreg}},
  \code{\link{extract.relevent}},
  \code{\link{extract.rms}},
  \code{\link{extract.robustbase}},
  \code{\link{extract.robust}},
  \code{\link{extract.sampleSelection}},
  \code{\link{extract.simex}}, \code{\link{extract.sna}},
  \code{\link{extract.spatialreg}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survey}},
  \code{\link{extract.survival}},
  \code{\link{extract.systemfit}},
  \code{\link{extract.tergm}}, \code{\link{extract.texreg}}
}
\author{
Philip Leifeld
}
\concept{extract}
