% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.feisr}
\alias{extract.feisr}
\alias{extract.feisr-methods}
\alias{extract,feis-method}
\alias{extract.feis}
\title{\code{\link{extract}} methods for models in the \pkg{feisr} package}
\usage{
\S4method{extract}{feis}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.rmse = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.groups}{Report the number of groups or alternatives?}

\item{include.rmse}{Report the root mean square error (RMSE; = residual
standard deviation) in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{feisr}
package, in particular:
\itemize{
  \item \code{feis} objects created by the \code{\link[feisr]{feis}}
    function in the \pkg{feisr} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.brglm}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.mfx}}, \code{\link{extract.mgcv}},
  \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.rms}},
  \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survival}}
}
\author{
Tobias Rüttenauer, Philip Leifeld
#' @examples
library("feisr")
library("texreg")
data("mwp", package = "feisr")
feis1.mod <- feis(lnw ~ marry | exp, data = mwp, id = "id")
feis2.mod <- feis(lnw ~ marry + enrol + as.factor(yeargr) | exp,
                  data = mwp,
                  id = "id")
sr1 <- screenreg(list(feis1.mod, feis2.mod), digits = 3)
tr1 <- texreg(list(feis1.mod, feis2.mod), digits = 3)
}
\concept{extract}
