% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.gamlss}
\alias{extract.gamlss}
\alias{extract.gamlss-methods}
\alias{extract,gamlss-method}
\title{\code{\link{extract}} methods for models in the \pkg{gamlss} package}
\usage{
\S4method{extract}{gamlss}(model, robust = FALSE, include.nobs = TRUE,
  include.nagelkerke = TRUE, include.gaic = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{robust}{If TRUE computes robust standard errors in the variance covariance matrix.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.nagelkerke}{Report Nagelkerke R^2 in the GOF block?}

\item{include.gaic}{Report Generalized Akaike's Information Criterion (AIC) in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{gamlss}
package, in particular:
\itemize{
  \item \code{gamlss} objects created by the \code{\link[gamlss]{gamlss}}
    function in the \pkg{gamlss} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,coeftest-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.MuMIn}},
  \code{\link{extract.RSiena}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.betareg}},
  \code{\link{extract.bife}}, \code{\link{extract.biglm}},
  \code{\link{extract.brglm}}, \code{\link{extract.brms}},
  \code{\link{extract.broom}},
  \code{\link{extract.btergm}},
  \code{\link{extract.censReg}},
  \code{\link{extract.dynlm}}, \code{\link{extract.erer}},
  \code{\link{extract.ergm}}, \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.lqmm}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.mhurdle}},
  \code{\link{extract.mlogit}},
  \code{\link{extract.mnlogit}},
  \code{\link{extract.nlme}}, \code{\link{extract.nnet}},
  \code{\link{extract.oglmx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.panelAR}},
  \code{\link{extract.pglm}}, \code{\link{extract.plm}},
  \code{\link{extract.rms}},
  \code{\link{extract.robustbase}},
  \code{\link{extract.robust}}, \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survey}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
