% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.h2o}
\alias{extract.h2o}
\alias{extract.h2o-methods}
\alias{extract,H2OBinomialModel-method}
\alias{extract.H2OBinomialModel}
\title{\code{\link{extract}} methods for models in the \pkg{h2o} package}
\usage{
\S4method{extract}{H2OBinomialModel}(model, standardized = FALSE,
  include.mse = TRUE, include.rsquared = TRUE,
  include.logloss = TRUE, include.meanerror = TRUE,
  include.auc = TRUE, include.gini = TRUE, include.deviance = TRUE,
  include.aic = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{standardized}{Report Standardized Coefficients in the Coefficients block?}

\item{include.mse}{Report the Mean Squared Error Value in the GOF block?}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.logloss}{Report the Log Loss Value in the GOF block?}

\item{include.meanerror}{Report the Mean Per-Class Error Value in the GOF block?}

\item{include.auc}{Report the Area Under Curve Score (AUC) in the GOF block?}

\item{include.gini}{Report the GINI Value in the GOF block?}

\item{include.deviance}{Report the deviance in the GOF block?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{h2o}
package, in particular:
\itemize{
  \item \code{H2OBinomialModel} objects created by the \code{\link[h2o]{h2o.glm}}
    function in the \pkg{h2o} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.brglm}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.mfx}}, \code{\link{extract.mgcv}},
  \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.rms}},
  \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
