% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.mfx}
\alias{extract.mfx}
\alias{extract.mfx-methods}
\alias{extract,betamfx-method}
\alias{extract.betamfx}
\alias{extract,betaor-method}
\alias{extract.betaor}
\alias{extract,logitmfx-method}
\alias{extract.logitmfx}
\alias{extract,probitmfx-method}
\alias{extract.probitmfx}
\alias{extract,logitor-method}
\alias{extract.logitor}
\alias{extract,negbinirr-method}
\alias{extract.negbinirr}
\alias{extract,poissonirr-method}
\alias{extract.poissonirr}
\alias{extract,negbinmfx-method}
\alias{extract.negbinmfx}
\alias{extract,poissonmfx-method}
\alias{extract.poissonmfx}
\title{\code{\link{extract}} methods for models in the \pkg{mfx} package}
\usage{
\S4method{extract}{betamfx}(model, include.pseudors = TRUE,
  include.loglik = TRUE, include.nobs = TRUE, ...)

\S4method{extract}{betaor}(model, include.pseudors = TRUE,
  include.loglik = TRUE, include.nobs = TRUE, ...)

\S4method{extract}{logitmfx}(model, include.nobs = TRUE,
  include.loglik = TRUE, include.deviance = TRUE, include.aic = TRUE,
  include.bic = TRUE, ...)

\S4method{extract}{probitmfx}(model, include.nobs = TRUE,
  include.loglik = TRUE, include.deviance = TRUE, include.aic = TRUE,
  include.bic = TRUE, ...)

\S4method{extract}{logitor}(model, include.nobs = TRUE,
  include.loglik = TRUE, include.deviance = TRUE, include.aic = TRUE,
  include.bic = TRUE, ...)

\S4method{extract}{negbinirr}(model, include.nobs = TRUE,
  include.loglik = TRUE, include.deviance = TRUE, include.aic = TRUE,
  include.bic = TRUE, ...)

\S4method{extract}{poissonirr}(model, include.nobs = TRUE,
  include.loglik = TRUE, include.deviance = TRUE, include.aic = TRUE,
  include.bic = TRUE, ...)

\S4method{extract}{negbinmfx}(model, include.nobs = TRUE,
  include.loglik = TRUE, include.deviance = TRUE, include.aic = TRUE,
  include.bic = TRUE, ...)

\S4method{extract}{poissonmfx}(model, include.nobs = TRUE,
  include.loglik = TRUE, include.deviance = TRUE, include.aic = TRUE,
  include.bic = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.pseudors}{Report pseudo R^2 in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.deviance}{Report the deviance in the GOF block?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{mfx}
package, in particular:
\itemize{
  \item \code{betamfx} objects created by the \code{\link[mfx]{betamfx}}
    function in the \pkg{mfx} package
  \item \code{betaor} objects created by the \code{\link[mfx]{betaor}}
    function in the \pkg{mfx} package
  \item \code{logitmfx} objects created by the \code{\link[mfx]{logitmfx}}
    function in the \pkg{mfx} package
  \item \code{logitor} objects created by the \code{\link[mfx]{logitor}}
    function in the \pkg{mfx} package
  \item \code{negbinirr} objects created by the \code{\link[mfx]{negbinirr}}
    function in the \pkg{mfx} package
  \item \code{negbinmfx} objects created by the \code{\link[mfx]{negbinmfx}}
    function in the \pkg{mfx} package
  \item \code{poissonirr} objects created by the \code{\link[mfx]{poissonirr}}
    function in the \pkg{mfx} package
  \item \code{poissonmfx} objects created by the \code{\link[mfx]{poissonmfx}}
    function in the \pkg{mfx} package
  \item \code{probitmfx} objects created by the \code{\link[mfx]{probitmfx}}
    function in the \pkg{mfx} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.brglm}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.mgcv}}, \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.rms}},
  \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
