% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.mgcv}
\alias{extract.mgcv}
\alias{extract.mgcv-methods}
\alias{extract,gam-method}
\alias{extract.gam}
\alias{extract,bam-method}
\alias{extract.bam}
\title{\code{\link{extract}} methods for models in the \pkg{mgcv} package}
\usage{
\S4method{extract}{gam}(model, include.smooth = TRUE,
  include.aic = TRUE, include.bic = TRUE, include.loglik = TRUE,
  include.deviance = TRUE, include.dev.expl = TRUE,
  include.dispersion = TRUE, include.rsquared = TRUE,
  include.gcv = TRUE, include.nobs = TRUE, include.nsmooth = TRUE,
  ...)

\S4method{extract}{bam}(model, include.smooth = TRUE,
  include.aic = TRUE, include.bic = TRUE, include.loglik = TRUE,
  include.deviance = TRUE, include.dev.expl = TRUE,
  include.dispersion = TRUE, include.rsquared = TRUE,
  include.gcv = TRUE, include.nobs = TRUE, include.nsmooth = TRUE,
  ...)
}
\arguments{
\item{model}{A model object.}

\item{include.smooth}{Report the smooth terms in the coefficient block?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.deviance}{Report the deviance in the GOF block?}

\item{include.dev.expl}{Report the deviance explained in the GOF block?}

\item{include.dispersion}{Report the dispersion in the GOF block?}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.gcv}{Report the minimised generalised cross-validation (GCV) in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.nsmooth}{Report number of the smooth terms in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{mgcv}
package, in particular:
\itemize{
  \item \code{bam} objects created by the \code{\link[mgcv]{bam}}
    function in the \pkg{mgcv} package
  \item \code{gam} objects created by the \code{\link[mgcv]{gam}}
    function in the \pkg{mgcv} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,H2OBinomialModel-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,gnm-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gmm}}, \code{\link{extract.lme4}},
  \code{\link{extract.mfx}}, \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
