% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.MuMIn}
\alias{extract.MuMIn}
\alias{extract.MuMIn-methods}
\alias{extract,averaging-method}
\alias{extract.averaging}
\alias{extract,model.selection-method}
\alias{extract.model.selection}
\title{\code{\link{extract}} methods for models in the \pkg{MuMIn} package}
\usage{
\S4method{extract}{averaging}(model, use.ci = FALSE,
  adjusted.se = FALSE, include.nobs = TRUE, ...)

\S4method{extract}{model.selection}(model, include.loglik = TRUE,
  include.aicc = TRUE, include.delta = TRUE, include.weight = TRUE,
  include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{use.ci}{Report confidence intervals in the GOF block?}

\item{adjusted.se}{Report adjusted standard error in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.aicc}{Report AICC in the GOF block?}

\item{include.delta}{Report the Dissimilarity index in the GOF block?}

\item{include.weight}{Report weight in the GOF block?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{MuMIn}
package, in particular:
\itemize{
  \item \code{averaging} objects created by the \code{\link[MuMIn]{model.avg}}
    function in the \pkg{MuMIn} package
  \item \code{model.selection} objects created by the \code{\link[MuMIn]{dredge}}
    function in the \pkg{MuMIn} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,coeftest-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.RSiena}}, \code{\link{extract.VGAM}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.betareg}},
  \code{\link{extract.bife}}, \code{\link{extract.biglm}},
  \code{\link{extract.brglm}}, \code{\link{extract.brms}},
  \code{\link{extract.broom}},
  \code{\link{extract.btergm}},
  \code{\link{extract.censReg}},
  \code{\link{extract.dynlm}}, \code{\link{extract.eha}},
  \code{\link{extract.erer}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.lqmm}},
  \code{\link{extract.metaSEM}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.mhurdle}},
  \code{\link{extract.mlogit}},
  \code{\link{extract.mnlogit}},
  \code{\link{extract.nlme}}, \code{\link{extract.nnet}},
  \code{\link{extract.oglmx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.panelAR}},
  \code{\link{extract.pglm}}, \code{\link{extract.plm}},
  \code{\link{extract.quantreg}},
  \code{\link{extract.relevent}},
  \code{\link{extract.rms}},
  \code{\link{extract.robustbase}},
  \code{\link{extract.robust}},
  \code{\link{extract.sampleSelection}},
  \code{\link{extract.simex}}, \code{\link{extract.sna}},
  \code{\link{extract.spatialreg}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survey}},
  \code{\link{extract.survival}},
  \code{\link{extract.systemfit}},
  \code{\link{extract.tergm}}, \code{\link{extract.texreg}}
}
\author{
Philip Leifeld
}
\concept{extract}
