% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.nlme}
\alias{extract.nlme}
\alias{extract.nlme-methods}
\alias{extract,gls-method}
\alias{extract.gls}
\alias{extract,gnls-method}
\alias{extract.gnls}
\alias{extract,lme-method}
\alias{extract.lme}
\title{\code{\link{extract}} methods for models in the \pkg{nlme} package}
\usage{
\S4method{extract}{gls}(model, include.aic = TRUE, include.bic = TRUE,
  include.loglik = TRUE, include.nobs = TRUE, ...)

\S4method{extract}{gnls}(model, include.aic = TRUE, include.bic = TRUE,
  include.loglik = TRUE, include.nobs = TRUE, ...)

\S4method{extract}{lme}(model, include.aic = TRUE, include.bic = TRUE,
  include.loglik = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.variance = FALSE, ...)

\method{extract}{nlme}(model, include.aic = TRUE, include.bic = TRUE,
  include.loglik = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.variance = FALSE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.groups}{Report groups in the GOF block?}

\item{include.variance}{Report variance in the GOF block?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{nlme}
package, in particular:
\itemize{
  \item \code{gls} objects created by the \code{\link[nlme]{gls}}
    function in the \pkg{nlme} package
  \item \code{gnls} objects created by the \code{\link[nlme]{gnls}}
    function in the \pkg{nlme} package
  \item \code{lme} objects created by the \code{\link[nlme]{lme}}
    function in the \pkg{nlme} package
  \item \code{nlme} objects created by the \code{\link[nlme]{nlme}}
    function in the \pkg{nlme} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,coeftest-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.MuMIn}},
  \code{\link{extract.RSiena}}, \code{\link{extract.VGAM}},
  \code{\link{extract.Zelig}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.betareg}},
  \code{\link{extract.bife}}, \code{\link{extract.biglm}},
  \code{\link{extract.brglm}}, \code{\link{extract.brms}},
  \code{\link{extract.broom}},
  \code{\link{extract.btergm}},
  \code{\link{extract.censReg}},
  \code{\link{extract.dynlm}}, \code{\link{extract.eha}},
  \code{\link{extract.erer}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.lqmm}},
  \code{\link{extract.metaSEM}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.mhurdle}},
  \code{\link{extract.mlogit}},
  \code{\link{extract.mnlogit}},
  \code{\link{extract.nnet}}, \code{\link{extract.oglmx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.panelAR}},
  \code{\link{extract.pglm}}, \code{\link{extract.plm}},
  \code{\link{extract.pscl}},
  \code{\link{extract.quantreg}},
  \code{\link{extract.relevent}},
  \code{\link{extract.rms}},
  \code{\link{extract.robustbase}},
  \code{\link{extract.robust}},
  \code{\link{extract.sampleSelection}},
  \code{\link{extract.simex}}, \code{\link{extract.sna}},
  \code{\link{extract.spatialreg}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survey}},
  \code{\link{extract.survival}},
  \code{\link{extract.systemfit}},
  \code{\link{extract.tergm}}, \code{\link{extract.texreg}}
}
\author{
Philip Leifeld
}
\concept{extract}
