% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.ordinal}
\alias{extract.ordinal}
\alias{extract.ordinal-methods}
\alias{extract,clm-method}
\alias{extract.clm}
\alias{extract,sclm-method}
\alias{extract.sclm}
\alias{extract,clmm-method}
\alias{extract.clmm}
\title{\code{\link{extract}} methods for models in the \pkg{ordinal} package}
\usage{
\S4method{extract}{clm}(model, include.thresholds = TRUE,
  include.aic = TRUE, include.bic = TRUE, include.loglik = TRUE,
  include.nobs = TRUE, ...)

\S4method{extract}{sclm}(model, include.thresholds = TRUE,
  include.aic = TRUE, include.bic = TRUE, include.loglik = TRUE,
  include.nobs = TRUE, ...)

\S4method{extract}{clmm}(model, include.thresholds = TRUE,
  include.loglik = TRUE, include.aic = TRUE, include.bic = TRUE,
  include.nobs = TRUE, include.groups = TRUE,
  include.variance = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.thresholds}{Report thresholds in the GOF block?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.groups}{Report groups in the GOF block?}

\item{include.variance}{Report variance in the GOF block?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{ordinal}
package, in particular:
\itemize{
  \item \code{clm} objects created by the \code{\link[ordinal]{clm}}
    function
  \item \code{sclm} objects
  \item \code{clmm} objects created by the \code{\link[ordinal]{clmm}}
    function
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,negbin-method}},
  \code{\link{extract,speedglm-method}},
  \code{\link{extract,speedlm-method}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.lme4}}, \code{\link{extract.mfx}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
