% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.rms}
\alias{extract.rms}
\alias{extract.rms-methods}
\alias{extract,lrm-method}
\alias{extract.lrm}
\alias{extract,ols-method}
\alias{extract.ols}
\title{\code{\link{extract}} methods for models in the \pkg{rms} package}
\usage{
\S4method{extract}{lrm}(model, include.pseudors = TRUE,
  include.lr = TRUE, include.nobs = TRUE, ...)

\S4method{extract}{ols}(model, include.nobs = TRUE,
  include.rsquared = TRUE, include.adjrs = TRUE,
  include.fstatistic = FALSE, include.lr = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.pseudors}{Report pseudo R^2 in the GOF block?}

\item{include.lr}{Report the model likelihood ration (L.R.) in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.fstatistic}{Report the F-statistic in the GOF block?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{rms}
package, in particular:
\itemize{
  \item \code{lrm} objects created by the \code{\link[rms]{lrm}}
    function in the \pkg{rms} package
  \item \code{ols} objects created by the \code{\link[rms]{ols}}
    function in the \pkg{rms} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.brglm}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.mfx}}, \code{\link{extract.mgcv}},
  \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survival}}
}
\author{
Fabrice Le Lec, Philip Leifeld
}
\concept{extract}
