% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.stats}
\alias{extract.stats}
\alias{extract.stats-methods}
\alias{extract,Arima-method}
\alias{extract.Arima}
\alias{extract,glm-method}
\alias{extract.glm}
\alias{extract,lm-method}
\alias{extract.lm}
\alias{extract,summary.lm-method}
\alias{extract.summary.lm}
\title{\code{\link{extract}} methods for models in the \pkg{stats} package}
\usage{
\S4method{extract}{Arima}(model, include.pvalues = FALSE,
  include.aic = TRUE, include.loglik = TRUE, ...)

\S4method{extract}{glm}(model, include.aic = TRUE, include.bic = TRUE,
  include.loglik = TRUE, include.deviance = TRUE,
  include.nobs = TRUE, ...)

\S4method{extract}{lm}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE,
  include.fstatistic = FALSE, include.rmse = FALSE, ...)

\S4method{extract}{summary.lm}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE,
  include.fstatistic = FALSE, include.rmse = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.pvalues}{Report p-values?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.deviance}{Report the deviance in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.fstatistic}{Report the F-statistic in the GOF block?}

\item{include.rmse}{Report the root mean square error (RMSE; = residual
standard deviation) in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.pvalues}{Report p-values?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{stats}
package, in particular:
\itemize{
  \item \code{Arima} objects created by the \code{\link[stats]{Arima}}
    function in the \pkg{stats} package
  \item \code{glm} objects created by the \code{\link[stats]{glm}}
    function in the \pkg{stats} package
  \item \code{lm} objects created by the \code{\link[stats]{lm}}
    function in the \pkg{stats} package
  \item \code{summary.lm} objects created by the \code{\link[stats]{summary.lm}}
    function in the \pkg{stats} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,coeftest-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.MuMIn}},
  \code{\link{extract.RSiena}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.betareg}},
  \code{\link{extract.bife}}, \code{\link{extract.biglm}},
  \code{\link{extract.brglm}}, \code{\link{extract.brms}},
  \code{\link{extract.broom}},
  \code{\link{extract.btergm}},
  \code{\link{extract.censReg}},
  \code{\link{extract.dynlm}}, \code{\link{extract.erer}},
  \code{\link{extract.ergm}}, \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.lqmm}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.mhurdle}},
  \code{\link{extract.mlogit}},
  \code{\link{extract.mnlogit}},
  \code{\link{extract.nlme}}, \code{\link{extract.nnet}},
  \code{\link{extract.oglmx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.panelAR}},
  \code{\link{extract.pglm}}, \code{\link{extract.plm}},
  \code{\link{extract.rms}},
  \code{\link{extract.robustbase}},
  \code{\link{extract.robust}}, \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.survey}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld, Claudia Zucca
}
\concept{extract}
