% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.systemfit}
\alias{extract.systemfit}
\alias{extract.systemfit-methods}
\alias{extract,systemfit-method}
\title{\code{\link{extract}} methods for models in the \pkg{systemfit} package}
\usage{
\S4method{extract}{systemfit}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE, beside = FALSE,
  include.suffix = FALSE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{beside}{if beside = TRUE show system elements in separate columns}

\item{include.suffix}{If include.suffix = TRUE show system elements in parethesis 
after each coefficient}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{systemfit}
package, in particular:
\itemize{
  \item \code{systemfit} objects created by the \code{\link[systemfit]{systemfit}}
    function in the \pkg{systemfit} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,coeftest-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.MuMIn}},
  \code{\link{extract.RSiena}}, \code{\link{extract.VGAM}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.betareg}},
  \code{\link{extract.bife}}, \code{\link{extract.biglm}},
  \code{\link{extract.brglm}}, \code{\link{extract.brms}},
  \code{\link{extract.broom}},
  \code{\link{extract.btergm}},
  \code{\link{extract.censReg}},
  \code{\link{extract.dynlm}}, \code{\link{extract.eha}},
  \code{\link{extract.erer}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.lqmm}},
  \code{\link{extract.metaSEM}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.mhurdle}},
  \code{\link{extract.mlogit}},
  \code{\link{extract.mnlogit}},
  \code{\link{extract.nlme}}, \code{\link{extract.nnet}},
  \code{\link{extract.oglmx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.panelAR}},
  \code{\link{extract.pglm}}, \code{\link{extract.plm}},
  \code{\link{extract.quantreg}},
  \code{\link{extract.relevent}},
  \code{\link{extract.rms}},
  \code{\link{extract.robustbase}},
  \code{\link{extract.robust}},
  \code{\link{extract.sampleSelection}},
  \code{\link{extract.simex}}, \code{\link{extract.sna}},
  \code{\link{extract.spatialreg}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survey}},
  \code{\link{extract.survival}},
  \code{\link{extract.tergm}}, \code{\link{extract.texreg}}
}
\author{
Philip Leifeld
}
\concept{extract}
