% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texreg.R
\name{get.data}
\alias{get.data}
\title{Extract all data necessary for generating a table from statistical models}
\usage{
get.data(l, ...)
}
\arguments{
\item{l}{A list of statistical models.}

\item{...}{Arguments to be passed over to the \code{\link{extract}} function,
such as \code{include.nobs} or \code{include.aic}. Details are provided in
the documentation of the \code{\link{extract}} function.}
}
\value{
A list of \linkS4class{texreg} objects.
}
\description{
Extract all data necessary for generating a table from a list of models.
}
\details{
This function applies the \code{link{extract}} function and its respective
methods to each element in a list of statistical models in order to extract
coefficients, standard errors, p-values, confidence intervals, and
goodness-of-fit statistics for generating a regression table.
}
\seealso{
\code{\link{extract}}
}
\author{
Philip Leifeld
}
\keyword{internal}
