% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{internal}
\alias{internal}
\alias{aggregate.matrix}
\alias{format.column}
\title{Internal functions for the \pkg{texreg} package}
\usage{
\method{aggregate}{matrix}(models, gof.names, custom.gof.names = NULL,
  custom.gof.rows = NULL, reorder.gof = NULL, digits = 2,
  leading.zero = TRUE, latex = FALSE, dcolumn = TRUE,
  returnobject = "m")

\method{format}{column}(x, single.row = FALSE, digits = 2)
}
\arguments{
\item{models}{A list of \code{texreg} objects, as created by the
\link{get.data} function.}

\item{gof.names}{A character vector of names for the GOF statistics, as
created by the \code{\link[=get.gof]{get.gof()}} function. This is needed to attach row
names to the GOF block or decimal matrix.}

\item{custom.gof.names}{A character vector of replacement names for the GOF
statistics. Must have the same length as \code{gof.names}. Can be
\code{NULL} if the original GOF names should be kept.}

\item{custom.gof.rows}{A named list of integer, numeric, or character vectors
that represent new rows to be inserted at the beginning of the GOF block.}

\item{reorder.gof}{A numeric vector with the desired reordered positions of
the GOF entries, for example \code{c(1, 2, 4, 3)} for swapping the last two
entries.}

\item{digits}{Number of decimal points (for formatting GOF values).}

\item{leading.zero}{Logical indicating whether leading zeroes should be used.}

\item{latex}{Is the output a LaTeX table? Needed for cell formatting.}

\item{dcolumn}{Is the \pkg{dcolumn} LaTeX package used for cell formatting?}

\item{returnobject}{This can be:
\describe{
\item{\code{"m"}}{for creating the coefficient block}
\item{\code{"gof.matrix"}}{for creating the GOF block}
}}

\item{x}{A character vector with coefficients (e.g., \code{"5.03 ***"}) and
either standard errors (e.g., \code{"(0.22)"}) or confidence intervals
(e.g., \code{"[-0.98; 0.24]"}).}

\item{single.row}{Are both the coefficient and the uncertainty measure in the
same row, as per the \code{single.row} argument of the table creation
functions?}

\item{digits}{The number of decimal places used to generate the table.}
}
\value{
A \code{matrix} object representing...
\itemize{
\item the coefficient block with all decimal places and with separate
columns for coefficients, standard errors and p-values, or confidence
intervals (for example, six columns for two models) or
\item the GOF block with one column per model with the first column
representing the GOF names and the remaining columns the formatted GOF
statistics as character objects.
}

A character vector of the same size as the input vector, but with
decimal-aligned values and equal character length of each entry.
}
\description{
Aggregate models and GOFs from \code{texreg} objects into a common matrix.

Format a column (given as vector) of the output matrix by adding spaces.
}
\details{
This function takes a list of \code{texreg} objects, as produced by the
\code{\link[=extract]{extract()}} function, and produces a matrix, either for the
coefficient block (\code{returnobject = "m"}) (in numeric form) or for the
goodness-of-fit (GOF) block of the table that is to be created
(\code{returnobject = "gof"}; in formatted character form). On the way, also
replace GOF names by custom GOF labels, then insert custom GOF rows if
necessary, and then reorder the GOF block if necessary.

This function accepts a vector of coefficients and either standard errors
(in parentheses) or confidence intervals (in square brackets, where the lower
and upper confidence interval are separated by a semicolon) and formats the
values nicely by aligning them at the decimal point. To do this, the function
adds necessary spaces both at the beginning and at the end. The vector to be
formatted by the function is usually a column from an output matrix as
generated by the \code{\link[=outputmatrix]{outputmatrix()}} function.
}
\seealso{
\code{\link[=extract]{extract()}}, \code{\link[=get.data]{get.data()}},
\code{\link[=get.gof]{get.gof()}}, \code{\link[=coeftostring]{coeftostring()}},
\code{\link[=replaceSymbols]{replaceSymbols()}}

\code{\link[=outputmatrix]{outputmatrix()}}, \code{\link[=matrixreg]{matrixreg()}}
}
\author{
Philip Leifeld

Philip Leifeld
}
\keyword{internal}
