% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texreg.R
\name{names2latex}
\alias{names2latex}
\title{Replace symbols in a character string or vector by LaTeX equivalents}
\usage{
names2latex(x)
}
\arguments{
\item{x}{A \code{character} object of arbitrary length.}
}
\value{
Same as the input object but with escaped or replaced symbols.
}
\description{
Replace symbols in a character string or vector by LaTeX equivalents-
}
\details{
This function is an internal helper function that takes a \code{character}
object or vector and replaces symbols like underscores, angle brackets, or
superscripted numbers by properly escaped LaTeX equivalents in order not to
break any LaTeX table code when the \code{link{texreg}} function is used.
}
\seealso{
\code{\link{texreg}}
}
\author{
Philip Leifeld
}
\keyword{internal}
