% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texreg.R
\name{print.texregTable}
\alias{print.texregTable}
\title{Prints a \code{texregTable} object.}
\usage{
\method{print}{texregTable}(x, ...)
}
\arguments{
\item{x}{A \code{texregTable} argument, as produced by \code{\link{texreg}}
and related functions.}

\item{...}{Additional arguments for the \code{\link[base]{cat}} function.}
}
\description{
Prints a \code{texregTable} object.
}
\author{
Philip Leifeld
}
