% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texreg.R
\name{reorder}
\alias{reorder}
\title{Reorder a matrix vertically according to a vector of new positions}
\usage{
reorder(mat, new.order)
}
\arguments{
\item{mat}{Input matrix.}

\item{new.order}{Vector of integer numbers with the new order of rows. The
new order must contain as many elements as the matrix has rows and it must
not contain NA values, duplicate entries, or gaps.}
}
\value{
Reordered matrix.
}
\description{
Reorder a matrix vertically according to a vector of new positions.
}
\details{
This function takes a matrix and reorders its rows based on a vector of new
positions.
}
\seealso{
\code{\link{matrixreg}}
}
\author{
Philip Leifeld
}
\keyword{internal}
