% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texreg.R
\name{show,texreg-method}
\alias{show,texreg-method}
\title{Show method for pretty output of \linkS4class{texreg} objects}
\usage{
\S4method{show}{texreg}(object)
}
\arguments{
\item{object}{The \linkS4class{texreg} object to display.}
}
\description{
Show method for pretty output of \linkS4class{texreg} objects.
}
\details{
Print the different slots of \linkS4class{texreg} objects to the screen.
}
\references{
Leifeld, Philip (2013). texreg: Conversion of Statistical Model
  Output in R to LaTeX and HTML Tables. Journal of Statistical Software
  55(8): 1-24. \url{http://www.jstatsoft.org/v55/i08/}.
}
\seealso{
\code{\link{extract}}, \code{\link{createTexreg}},
  \code{\link{screenreg}}
}
\author{
Philip Leifeld
}
