% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texreg.R
\name{texreg-package}
\alias{texreg-package}
\title{Conversion of \R Regression Output to LaTeX or HTML Tables}
\description{
\pkg{texreg} converts coefficients, standard errors, uncertainty measures,
and goodness-of-fit statistics of statistical models into LaTeX or HTML
tables or into nicely formatted screen output for the \R console. A list of
several models can be combined in a single table. The output is
customizable. New model types can be easily implemented. Confidence
intervals can be used instead of standard errors and p-values.
}
\references{
Leifeld, Philip (2013). texreg: Conversion of Statistical Model
  Output in R to LaTeX and HTML Tables. Journal of Statistical Software
  55(8): 1-24. \doi{10.18637/jss.v055.i08}.
}
\seealso{
\code{\link{extract}} \code{\link{texreg}}
}
\author{
Philip Leifeld
}
\keyword{package}
