% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,apollo-method}
\alias{extract,apollo-method}
\alias{extract.apollo}
\title{\code{\link{extract}} method for \code{apollo} objects}
\usage{
\S4method{extract}{apollo}(
  model,
  wtpest = NULL,
  se = "rob",
  include.loglik = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{An object of class \code{apollo}.}

\item{wtpest}{Optional \code{data.frame} of willingness-to-pay estimates
(and standard errors).}

\item{se}{Which standard errors to use: \code{"rob"}, \code{"normal"}, or
\code{"bs"}.}

\item{include.loglik}{Report the log likelihood (including for the null
model) in the GOF block?}

\item{include.nobs}{Report the number of observations and individuals in the
GOF block?}

\item{...}{Currently ignored.}
}
\description{
\code{\link{extract}} method for \code{apollo} objects created by
\code{apollo::apollo_estimate()}.
}
\author{
Julian Sagebiel \email{julian.sagebiel@idiv.de}, Philip Leifeld
}
