% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,Arima-method}
\alias{extract,Arima-method}
\alias{extract.Arima}
\title{\code{\link{extract}} method for \code{Arima} objects}
\usage{
\S4method{extract}{Arima}(model, include.pvalues = FALSE,
  include.aic = TRUE, include.loglik = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.pvalues}{Report p-values?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{Arima} objects. These objects are
created by the \code{\link[stats]{arima}} function in the \pkg{stats} package.
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract.MASS}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gmm}}, \code{\link{extract.lme4}},
  \code{\link{extract.mfx}}, \code{\link{extract.mgcv}},
  \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
