% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,averaging-method}
\alias{extract,averaging-method}
\alias{extract.averaging}
\title{\code{\link{extract}} method for \code{averaging} objects}
\usage{
\S4method{extract}{averaging}(model, use.ci = FALSE, adjusted.se = FALSE, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{use.ci}{Report confidence intervals in the GOF block?}

\item{adjusted.se}{Report adjusted standard error in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{averaging} objects created by the
\code{\link[MuMIn]{model.avg}} function in the \pkg{MuMIn} package.
}
