% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,betareg-method}
\alias{extract,betareg-method}
\alias{extract.betareg}
\title{\code{\link{extract}} method for \code{betareg} objects}
\usage{
\S4method{extract}{betareg}(
  model,
  include.precision = TRUE,
  include.pseudors = TRUE,
  include.loglik = TRUE,
  include.nobs = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.precision}{Report precision in the GOF block?}

\item{include.pseudors}{Report pseudo R^2 in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{betareg} objects created by the
\code{\link[betareg]{betareg}} function in the \pkg{betareg} package.
}
