% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,brmsfit-method}
\alias{extract,brmsfit-method}
\alias{extract.brmsfit}
\title{\code{\link{extract}} method for \code{brmsfit} objects}
\usage{
\S4method{extract}{brmsfit}(model, use.HDI = TRUE, level = 0.9,
  include.random = TRUE, include.rsquared = TRUE,
  include.nobs = TRUE, include.loo.ic = TRUE, reloo = FALSE,
  include.waic = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{use.HDI}{Report highest posterior density (HPD) intervals (HDI) using
the \code{\link[coda]{HPDinterval}} function in the \pkg{coda} package,
with the probability given in the \code{level} argument, instead of the
default 95 percent posterior quantiles?}

\item{level}{Significance level (\code{1 - alpha}) for the \code{use.HDI}
argument.}

\item{include.random}{Include random effects (standard deviations) in the GOF
block of the table?}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.loo.ic}{Report Leave-One-Out Information Criterion?}

\item{reloo}{Recompute exact cross-validation for problematic observations
for which approximate leave-one-out cross-validation may return incorrect
results? This is done using the \code{\link[brms]{reloo}} function and may
take some time to compute.}

\item{include.waic}{Report Widely Applicable Information Criterion (WAIC)?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{brmsfit} objects. These objects are
created by the \code{\link[brms]{brm}} function in the \pkg{brms} package.
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,H2OBinomialModel-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,gnm-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gmm}}, \code{\link{extract.lme4}},
  \code{\link{extract.mfx}}, \code{\link{extract.mgcv}},
  \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.survival}}
}
\author{
Hyunjin (Jin) Song, Philip Leifeld
}
\concept{extract}
