% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,coeftest-method}
\alias{extract,coeftest-method}
\alias{extract.coeftest}
\title{\code{\link{extract}} method for \code{coeftest} objects}
\usage{
\S4method{extract}{coeftest}(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{coeftest} objects. These objects are
created by the \code{\link[lmtest]{coeftest}} function in the \pkg{lmtest}
package.
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,gel-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,negbin-method}},
  \code{\link{extract,speedglm-method}},
  \code{\link{extract,speedlm-method}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.lme4}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
