% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,feis-method}
\alias{extract,feis-method}
\alias{extract.feis}
\title{\code{\link{extract}} method for \code{feis} objects}
\usage{
\S4method{extract}{feis}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.rmse = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.rmse}{Report the root mean square error (RMSE; = residual
standard deviation) in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function
for \code{feis} objects.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{feis} objects. These objects are
created by the \code{\link[feisr]{feis}} function in the \pkg{feisr} package.
}
\examples{
library("feisr")
library("texreg")
data("mwp", package = "feisr")
feis1.mod <- feis(lnw ~ marry | exp, data = mwp, id = "id")
feis2.mod <- feis(lnw ~ marry + enrol + as.factor(yeargr) | exp,
                  data = mwp,
                  id = "id")
sr1 <- screenreg(list(feis1.mod, feis2.mod), digits = 3)
tr1 <- texreg(list(feis1.mod, feis2.mod), digits = 3)

}
\seealso{
\link{extract}

Other extract: \code{\link{extract,dynlm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,lme4-method}},
  \code{\link{extract,speedlm-method}}
}
\author{
Tobias Rüttenauer, Philip Leifeld
}
\concept{extract}
