% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,felm-method}
\alias{extract,felm-method}
\alias{extract.felm}
\title{\code{\link{extract}} method for \code{felm} objects}
\usage{
\S4method{extract}{felm}(model, include.nobs = TRUE,
  include.rsquared = TRUE, include.adjrs = TRUE,
  include.fstatistic = FALSE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.fstatistic}{Report the F-statistic in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{felm} objects. These objects are
created by the \code{\link[lfe]{felm}} function in the \pkg{lfe} package.
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,negbin-method}},
  \code{\link{extract,speedglm-method}},
  \code{\link{extract,speedlm-method}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.lme4}}, \code{\link{extract.mfx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.survival}}
}
\author{
Christoph Riedl, Claudia Zucca, Philip Leifeld
}
\concept{extract}
