% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,fixest-method}
\alias{extract,fixest-method}
\alias{extract.fixest}
\title{\code{\link{extract}} method for \code{fixest} objects}
\usage{
\S4method{extract}{fixest}(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the extract methods for individual model types (e.g.
\code{\link{extract.fixest.feols}}).}
}
\description{
\code{\link{extract}} method for model objects created by the
model fitting functions in the \pkg{fixest} package.
}
\author{
Christopher Poliquin
}
