% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,gee-method}
\alias{extract,gee-method}
\alias{extract.gee}
\title{\code{\link{extract}} method for \code{gee} objects}
\usage{
\S4method{extract}{gee}(model, robust = TRUE, include.scale = TRUE, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{robust}{If TRUE computes robust standard errors in the
variance-covariance matrix.}

\item{include.scale}{Report the dispersion or scale parameter?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}
}
\description{
\code{\link{extract}} method for \code{gee} objects created by the
\code{\link[gee]{gee}} function in the \pkg{gee} package.
}
