% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract,ivreg-method}
\alias{extract,ivreg-method}
\alias{extract.ivreg}
\title{\code{\link{extract}} method for \code{ivreg} objects}
\usage{
\S4method{extract}{ivreg}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE,
  include.fstatistic = FALSE, include.rmse = FALSE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.fstatistic}{Report the F-statistic in the GOF block?}

\item{include.rmse}{Report the root mean square error (RMSE; = residual
standard deviation) in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function
for \code{lm} objects.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} method for \code{ivreg} objects. These objects are
created by the \code{\link[AER]{ivreg}} function in the \pkg{AER}
package.
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,dynlm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,lme4-method}},
  \code{\link{extract,speedlm-method}}
}
\author{
Philip Leifeld
}
\concept{extract}
