% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,lmrob-method}
\alias{extract,lmrob-method}
\alias{extract.lmrob}
\alias{extract,lmRob-method}
\alias{extract.lmRob}
\title{\code{\link{extract}} method for \code{lmrob} objects}
\usage{
\S4method{extract}{lmrob}(model, include.nobs = TRUE, ...)

\S4method{extract}{lmRob}(
  model,
  include.rsquared = TRUE,
  include.nobs = TRUE,
  include.rmse = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines, in this
case to the \code{summary} method for the object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.rmse}{Report the root mean square error (RMSE; = residual
standard deviation) in the GOF block?}
}
\description{
\code{\link{extract}} method for \code{lmrob} objects created by the
\code{\link[robustbase]{lmrob}} function in the \pkg{robustbase} package.

\code{\link{extract}} method for \code{lmRob} objects created by the
\code{\link[robust]{lmRob}} function in the \pkg{robust} package.
}
