% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,logitr-method}
\alias{extract,logitr-method}
\alias{logitr}
\title{\code{\link{extract}} method for \code{logitr} objects}
\usage{
\S4method{extract}{logitr}(
  model,
  include.nobs = TRUE,
  include.loglik = TRUE,
  include.aic = TRUE,
  include.bic = TRUE,
  ...
)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Include the number of observations in summary table?}

\item{include.loglik}{Include the log-likelihood in summary table?}

\item{include.aic}{Include the the AIC in summary table?}

\item{include.bic}{Include the the BIC in summary table?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{logitr} objects created by the
\code{logitr} function in the \pkg{logitr} package.
}
\examples{
library("logitr")
library("texreg")

# Estimate a preference space model
mnl_pref <- logitr(
  data    = yogurt,
  outcome = "choice",
  obsID   = "obsID",
  pars    = c("price", "feat", "brand")
)

# Print summary table to screen
screenreg(mnl_pref, stars = c(0.01, 0.05, 0.1))

# Print LaTeX code for summary table to screen
texreg(mnl_pref, stars = c(0.01, 0.05, 0.1))
}
