% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract,wls-method}
\alias{extract,wls-method}
\alias{extract.wls}
\title{\code{\link{extract}} method for \code{wls} objects}
\usage{
\S4method{extract}{wls}(model, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Custom parameters, which are handed over to subroutines. Currently
not in use.}
}
\description{
\code{\link{extract}} method for \code{wls} objects created by the
\code{\link[metaSEM]{wls}} function in the \pkg{metaSEM} package.
}
\author{
Christoph Riedl <c.riedl@neu.edu>
}
