% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.AER}
\alias{extract.AER}
\alias{extract.AER-methods}
\alias{extract,ivreg-method}
\alias{extract.ivreg}
\alias{extract,tobit-method}
\alias{extract.tobit}
\title{\code{\link{extract}} methods for models in the \pkg{AER} package}
\usage{
\S4method{extract}{ivreg}(model, include.rsquared = TRUE,
  include.adjrs = TRUE, include.nobs = TRUE,
  include.fstatistic = FALSE, include.rmse = FALSE, ...)

\S4method{extract}{tobit}(model, include.aic = TRUE,
  include.bic = TRUE, include.loglik = TRUE, include.deviance = TRUE,
  include.nobs = FALSE, include.censnobs = TRUE, include.wald = TRUE,
  ...)
}
\arguments{
\item{model}{A model object.}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.adjrs}{Report adjusted R^2 in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.fstatistic}{Report the F-statistic in the GOF block?}

\item{include.rmse}{Report the root mean square error (RMSE; = residual
standard deviation) in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.deviance}{Report the deviance in the GOF block?}

\item{include.censnobs}{Report the Censored number of observation parameters in the GOF block?}

\item{include.wald}{Report the Wald Test Value in the GOF block?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{AER}
package, in particular:
\itemize{
  \item \code{ivreg} objects created by the \code{\link[AER]{ivreg}}
    function in the \pkg{AER} package
  \item \code{tobit} objects created by the \code{\link[AER]{tobit}}
    function in the \pkg{AER} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,H2OBinomialModel-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,gnm-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.MASS}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gmm}}, \code{\link{extract.lme4}},
  \code{\link{extract.mfx}}, \code{\link{extract.mgcv}},
  \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
