% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.forecast}
\alias{extract.forecast}
\alias{extract.forecast-methods}
\alias{extract,ARIMA-method}
\alias{extract.ARIMA}
\alias{extract,ets-method}
\alias{extract.ets}
\alias{extract,forecast-method}
\title{\code{\link{extract}} methods for models in the \pkg{forecast} package}
\usage{
\S4method{extract}{ARIMA}(model, include.pvalues = FALSE,
  include.aic = TRUE, include.aicc = TRUE, include.bic = TRUE,
  include.loglik = TRUE, ...)

\S4method{extract}{ets}(model, include.pvalues = FALSE,
  include.aic = TRUE, include.aicc = TRUE, include.bic = TRUE,
  include.loglik = TRUE, ...)

\S4method{extract}{forecast}(model, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.pvalues}{Report p-values?}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.aicc}{Report AICC in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for models in the \pkg{forecast} package.
}
\details{
\code{\link{extract}} methods for objects in the \pkg{forecast} package, in
particular:
\itemize{
  \item \code{ARIMA} objects, which are created by the
    \code{\link[forecast]{Arima}} function
  \item \code{ets} objects, which are created by the
    \code{\link[forecast]{ets}} function
  \item \code{forecast} objects, which are created by several functions in
    the \pkg{forecast} package, including the
    \code{\link[forecast]{forecast}} function and the
    \code{\link[forecast]{holt}} function
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,H2OBinomialModel-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,gnm-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,lmRob-method}},
  \code{\link{extract,lqmm-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MASS}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}}, \code{\link{extract.gmm}},
  \code{\link{extract.lme4}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}}, \code{\link{extract.nlme}},
  \code{\link{extract.ordinal}}, \code{\link{extract.sna}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
