% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.lme4}
\alias{extract.lme4}
\alias{extract.lme4-methods}
\alias{extract,lme4-method}
\alias{extract,mer-method}
\alias{extract.merMod}
\alias{extract.mer}
\alias{extract,lmerMod-method}
\alias{extract.lmerMod}
\alias{extract.lmer}
\alias{extract,glmerMod-method}
\alias{extract.glmerMod}
\alias{extract.glmer}
\alias{extract,nlmerMod-method}
\alias{extract.nlmerMod}
\alias{extract.nlmer}
\title{\code{\link{extract}} methods for models in the \pkg{lme4} package}
\usage{
\S4method{extract}{lme4}(model, method = c("naive", "profile", "boot",
  "Wald"), level = 0.95, nsim = 1000, include.aic = TRUE,
  include.bic = TRUE, include.dic = FALSE, include.deviance = FALSE,
  include.loglik = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.variance = TRUE, ...)

\S4method{extract}{mer}(model, method = c("naive", "profile", "boot",
  "Wald"), level = 0.95, nsim = 1000, include.aic = TRUE,
  include.bic = TRUE, include.dic = FALSE, include.deviance = FALSE,
  include.loglik = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.variance = TRUE, ...)

\S4method{extract}{lmerMod}(model, method = c("naive", "profile", "boot",
  "Wald"), level = 0.95, nsim = 1000, include.aic = TRUE,
  include.bic = TRUE, include.dic = FALSE, include.deviance = FALSE,
  include.loglik = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.variance = TRUE, ...)

\S4method{extract}{glmerMod}(model, method = c("naive", "profile",
  "boot", "Wald"), level = 0.95, nsim = 1000, include.aic = TRUE,
  include.bic = TRUE, include.dic = FALSE, include.deviance = FALSE,
  include.loglik = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.variance = TRUE, ...)

\S4method{extract}{nlmerMod}(model, method = c("naive", "profile",
  "boot", "Wald"), level = 0.95, nsim = 1000, include.aic = TRUE,
  include.bic = TRUE, include.dic = FALSE, include.deviance = FALSE,
  include.loglik = TRUE, include.nobs = TRUE, include.groups = TRUE,
  include.variance = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{method}{The method used to compute confidence intervals or p-values.
The default value \code{"naive"} computes naive p-values while the other
methods compute confidence intervals using the \code{confint} function. See
\code{\link[lme4]{confint.merMod}}.}

\item{level}{Confidence level (\code{1 - alpha}) for computing confidence
intervals.}

\item{nsim}{In (generalized) linear mixed effects models: the MCMC sample
size or number of bootstrapping replications on the basis of which
confidence intervals are computed (only if the \code{method} argument does
not specify \code{"naive"}, which is the default behavior). Note: large
values may take considerable computing time.}

\item{include.aic}{Report Akaike's information criterion (AIC)?}

\item{include.bic}{Report the Bayesian information criterion (BIC)?}

\item{include.dic}{Report the deviance information criterion (DIC)?}

\item{include.deviance}{Report the deviance?}

\item{include.loglik}{Report the log-likelihood?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.groups}{Report the number of groups or alternatives?}

\item{include.variance}{Report group variances?}

\item{...}{Additional arguments for the \code{\link[lme4]{fixef}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for \code{merMod} objects in the \pkg{lme4}
package, in particular:
\itemize{
  \item \code{lmerMod} objects created by the \code{\link[lme4]{lmer}}
    function
  \item \code{glmerMod} objects created by the \code{\link[lme4]{glmer}}
    function
  \item \code{nlmerMod} objects created by the \code{\link[lme4]{nlmer}}
    function
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,gee-method}},
  \code{\link{extract,geeglm-method}},
  \code{\link{extract,gel-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,negbin-method}},
  \code{\link{extract,speedglm-method}},
  \code{\link{extract,speedlm-method}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.mfx}}, \code{\link{extract.mgcv}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
