% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.MASS}
\alias{extract.MASS}
\alias{extract.MASS-methods}
\alias{extract,negbin-method}
\alias{extract.negbin}
\alias{extract,polr-method}
\alias{extract.polr}
\alias{extract,rlm-method}
\alias{extract.rlm}
\alias{extract,glmmPQL-method}
\alias{extract.glmmPQL}
\title{\code{\link{extract}} methods for models in the \pkg{MASS} package}
\usage{
\S4method{extract}{negbin}(model, include.aic = TRUE,
  include.bic = TRUE, include.loglik = TRUE, include.deviance = TRUE,
  include.nobs = TRUE, ...)

\S4method{extract}{polr}(model, include.thresholds = FALSE,
  include.aic = TRUE, include.bic = TRUE, include.loglik = TRUE,
  include.deviance = TRUE, include.nobs = TRUE, ...)

\S4method{extract}{rlm}(model, include.nobs = TRUE, ...)

\S4method{extract}{glmmPQL}(model, include.aic = TRUE,
  include.bic = TRUE, include.loglik = TRUE, include.nobs = TRUE,
  include.groups = TRUE, include.variance = FALSE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.deviance}{Report the deviance in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.thresholds}{Report thresholds in the GOF block?}

\item{include.groups}{Report the number of groups or alternatives?}

\item{include.variance}{Report group variances?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{MASS}
package, in particular:
\itemize{
  \item \code{glmmPQL} objects created by the \code{\link[MASS]{glmmPQL}}
    function in the \pkg{MASS} package
  \item \code{negbin} objects created by the \code{\link[MASS]{negbin}}
    function in the \pkg{MASS} package
  \item \code{polr} objects created by the \code{\link[MASS]{polr}}
    function in the \pkg{MASS} package
  \item \code{rlm} objects created by the \code{\link[MASS]{rlm}}
    function in the \pkg{MASS} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,coeftest-method}},
  \code{\link{extract.AER}}, \code{\link{extract.MuMIn}},
  \code{\link{extract.RSiena}}, \code{\link{extract.VGAM}},
  \code{\link{extract.alpaca}},
  \code{\link{extract.betareg}},
  \code{\link{extract.bife}}, \code{\link{extract.biglm}},
  \code{\link{extract.brglm}}, \code{\link{extract.brms}},
  \code{\link{extract.broom}},
  \code{\link{extract.btergm}},
  \code{\link{extract.censReg}},
  \code{\link{extract.dynlm}}, \code{\link{extract.eha}},
  \code{\link{extract.erer}}, \code{\link{extract.ergm}},
  \code{\link{extract.fGarch}},
  \code{\link{extract.feisr}},
  \code{\link{extract.forecast}},
  \code{\link{extract.gamlss}},
  \code{\link{extract.geepack}}, \code{\link{extract.gee}},
  \code{\link{extract.gmm}}, \code{\link{extract.gnm}},
  \code{\link{extract.h2o}},
  \code{\link{extract.latentnet}},
  \code{\link{extract.lfe}}, \code{\link{extract.lme4}},
  \code{\link{extract.lqmm}},
  \code{\link{extract.metaSEM}}, \code{\link{extract.mfx}},
  \code{\link{extract.mgcv}},
  \code{\link{extract.mhurdle}},
  \code{\link{extract.mlogit}},
  \code{\link{extract.mnlogit}},
  \code{\link{extract.nlme}}, \code{\link{extract.nnet}},
  \code{\link{extract.oglmx}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.panelAR}},
  \code{\link{extract.pglm}}, \code{\link{extract.plm}},
  \code{\link{extract.quantreg}},
  \code{\link{extract.relevent}},
  \code{\link{extract.rms}},
  \code{\link{extract.robustbase}},
  \code{\link{extract.robust}},
  \code{\link{extract.sampleSelection}},
  \code{\link{extract.simex}}, \code{\link{extract.sna}},
  \code{\link{extract.spatialreg}},
  \code{\link{extract.speedglm}},
  \code{\link{extract.stats}},
  \code{\link{extract.survey}},
  \code{\link{extract.survival}},
  \code{\link{extract.systemfit}},
  \code{\link{extract.tergm}}, \code{\link{extract.texreg}}
}
\author{
Philip Leifeld
}
\concept{extract}
