% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.mfx}
\alias{extract.mfx}
\alias{extract.mfx-methods}
\alias{extract,betamfx-method}
\alias{extract.betamfx}
\alias{extract,betaor-method}
\alias{extract.betaor}
\title{\code{\link{extract}} methods for models in the \pkg{mfx} package}
\usage{
\S4method{extract}{betamfx}(model, include.pseudors = TRUE,
  include.loglik = TRUE, include.nobs = TRUE, ...)

\S4method{extract}{betaor}(model, include.pseudors = TRUE,
  include.loglik = TRUE, include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.pseudors}{Report pseudo R^2 in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models in the \pkg{mfx}
package, in particular:
\itemize{
  \item \code{betamfx} objects created by the \code{\link[mfx]{betamfx}}
    function in the \pkg{mfx} package
  \item \code{betaor} objects created by the \code{\link[mfx]{betaor}}
    function in the \pkg{mfx} package
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,ARIMA-method}},
  \code{\link{extract,Arima-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,negbin-method}},
  \code{\link{extract,speedglm-method}},
  \code{\link{extract,speedlm-method}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.lme4}},
  \code{\link{extract.ordinal}},
  \code{\link{extract.survival}}
}
\author{
Philip Leifeld
}
\concept{extract}
