% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract details from statistical models for table construction}
\usage{
extract(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines. The
arguments are usually passed to the \code{summary} function, but in some
cases to other functions, such as the \code{\link[lme4]{fixef}} function in
the \pkg{lme4} package.}
}
\value{
The function returns a \linkS4class{texreg} object.
}
\description{
Extract details from statistical models for table construction. The function
has methods for a range of statistical models, which are listed at the bottom
of this help page.
}
\details{
The \code{\link{extract}} function serves to retrieve coefficients, standard
errors, p-values, confidence intervals, and goodness-of-fit statistics from
statistical models in \R. More than 100 \code{\link{extract}} methods
("extensions") for various statistical models are available. The function
returns a \linkS4class{texreg} object.

\code{\link{extract}} is a generic function, which extracts coefficients and
GOF measures from statistical model objects. \code{\link{extract}} methods
for the specific model types are called by the generic \code{\link{extract}}
function if it encounters a model known to be handled by the specific method.
The output is a \linkS4class{texreg} object, which is subsequently used by
the \code{\link{texreg}} function and related functions.

Users can contribute their own extensions for additional statistical models.
Examples are contained in the article in the Journal of Statistical Software
referenced below. Suggestions can be submitted as pull requests at
\url{https://github.com/leifeld/texreg/pulls} or through the issue tracker at
\url{https://github.com/leifeld/texreg/issues}.
}
\references{
Leifeld, Philip (2013). texreg: Conversion of Statistical Model
  Output in R to LaTeX and HTML Tables. Journal of Statistical Software
  55(8): 1-24. \url{http://www.jstatsoft.org/v55/i08/}.
}
\seealso{
\code{\link{createTexreg}}, \code{\link{matrixreg}},
  \code{\link{screenreg}}, \code{\link{texreg}}

Other extract: 
\code{\link{extract,ANY-method}},
\code{\link{extract,ARIMA-method}},
\code{\link{extract,Arima-method}},
\code{\link{extract,H2OBinomialModel-method}},
\code{\link{extract,Zelig-method}},
\code{\link{extract,aftreg-method}},
\code{\link{extract,averaging-method}},
\code{\link{extract,bam-method}},
\code{\link{extract,betamfx-method}},
\code{\link{extract,betaor-method}},
\code{\link{extract,betareg-method}},
\code{\link{extract,bife-method}},
\code{\link{extract,biglm-method}},
\code{\link{extract,brglm-method}},
\code{\link{extract,brmsfit-method}},
\code{\link{extract,btergm-method}},
\code{\link{extract,censReg-method}},
\code{\link{extract,clm-method}},
\code{\link{extract,clmm-method}},
\code{\link{extract,clogit-method}},
\code{\link{extract,coeftest-method}},
\code{\link{extract,coxph-method}},
\code{\link{extract,coxph.penal-method}},
\code{\link{extract,coxreg-method}},
\code{\link{extract,dynlm-method}},
\code{\link{extract,ergm-method}},
\code{\link{extract,ergmm-method}},
\code{\link{extract,ets-method}},
\code{\link{extract,fGARCH-method}},
\code{\link{extract,feglm-method}},
\code{\link{extract,feis-method}},
\code{\link{extract,felm-method}},
\code{\link{extract,forecast-method}},
\code{\link{extract,gam-method}},
\code{\link{extract,gamlss-method}},
\code{\link{extract,gee-method}},
\code{\link{extract,geeglm-method}},
\code{\link{extract,gel-method}},
\code{\link{extract,glm-method}},
\code{\link{extract,glmerMod-method}},
\code{\link{extract,glmmPQL-method}},
\code{\link{extract,glmmadmb-method}},
\code{\link{extract,glmrob-method}},
\code{\link{extract,gls-method}},
\code{\link{extract,gmm-method}},
\code{\link{extract,gnls-method}},
\code{\link{extract,gnm-method}},
\code{\link{extract,hurdle-method}},
\code{\link{extract,lm-method}},
\code{\link{extract,lmRob-method}},
\code{\link{extract,lme-method}},
\code{\link{extract,lme4-method}},
\code{\link{extract,lmerMod-method}},
\code{\link{extract,lmrob-method}},
\code{\link{extract,lnam-method}},
\code{\link{extract,logitmfx-method}},
\code{\link{extract,logitor-method}},
\code{\link{extract,lqmm-method}},
\code{\link{extract,lrm-method}},
\code{\link{extract,maBina-method}},
\code{\link{extract,merMod-method}},
\code{\link{extract,mhurdle-method}},
\code{\link{extract,mlogit-method}},
\code{\link{extract,mnlogit-method}},
\code{\link{extract,model.selection-method}},
\code{\link{extract,mtergm-method}},
\code{\link{extract,multinom-method}},
\code{\link{extract,negbin-method}},
\code{\link{extract,negbinirr-method}},
\code{\link{extract,negbinmfx-method}},
\code{\link{extract,netlogit-method}},
\code{\link{extract,nlme-method}},
\code{\link{extract,nlmerMod-method}},
\code{\link{extract,oglmx-method}},
\code{\link{extract,ols-method}},
\code{\link{extract,panelAR-method}},
\code{\link{extract,pglm-method}},
\code{\link{extract,pgmm-method}},
\code{\link{extract,phreg-method}},
\code{\link{extract,plm-method}},
\code{\link{extract,pmg-method}},
\code{\link{extract,poissonirr-method}},
\code{\link{extract,poissonmfx-method}},
\code{\link{extract,polr-method}},
\code{\link{extract,probitmfx-method}},
\code{\link{extract,rem.dyad-method}},
\code{\link{extract,rlm-method}},
\code{\link{extract,rq-method}},
\code{\link{extract,sarlm-method}},
\code{\link{extract,sclm-method}},
\code{\link{extract,selection-method}},
\code{\link{extract,sienaFit-method}},
\code{\link{extract,simex-method}},
\code{\link{extract,speedglm-method}},
\code{\link{extract,speedlm-method}},
\code{\link{extract,stergm-method}},
\code{\link{extract,summary.lm-method}},
\code{\link{extract,survreg-method}},
\code{\link{extract,survreg.penal-method}},
\code{\link{extract,svyglm-method}},
\code{\link{extract,systemfit-method}},
\code{\link{extract,texreg-method}},
\code{\link{extract,tobit-method}},
\code{\link{extract,vglm-method}},
\code{\link{extract,weibreg-method}},
\code{\link{extract,wls-method}},
\code{\link{extract,zelig-method}},
\code{\link{extract,zeroinfl-method}}
}
\author{
Philip Leifeld
}
\concept{extract}
