% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\docType{methods}
\name{extract.survival}
\alias{extract.survival}
\alias{extract.survival-methods}
\alias{extract,clogit-method}
\alias{extract.clogit}
\alias{extract,coxph-method}
\alias{extract.coxph.penal}
\alias{extract,coxph.penal-method}
\alias{extract,survreg-method}
\alias{extract.survreg}
\alias{extract,survreg.penal-method}
\alias{extract.survreg.penal}
\title{\code{\link{extract}} method for models in the \pkg{survival} package}
\usage{
\S4method{extract}{clogit}(model, include.aic = TRUE,
  include.rsquared = TRUE, include.maxrs = TRUE,
  include.events = TRUE, include.nobs = TRUE,
  include.missings = TRUE, ...)

\S4method{extract}{coxph}(model, include.aic = TRUE,
  include.rsquared = TRUE, include.maxrs = TRUE,
  include.events = TRUE, include.nobs = TRUE,
  include.missings = TRUE, include.zph = TRUE, ...)

\S4method{extract}{coxph.penal}(model, include.aic = TRUE,
  include.rsquared = TRUE, include.maxrs = TRUE,
  include.events = TRUE, include.nobs = TRUE,
  include.missings = TRUE, include.zph = TRUE, ...)

\S4method{extract}{survreg}(model, include.aic = TRUE,
  include.bic = TRUE, include.loglik = TRUE, include.deviance = TRUE,
  include.nobs = TRUE, ...)

\S4method{extract}{survreg.penal}(model, include.aic = TRUE,
  include.bic = TRUE, include.loglik = TRUE, include.deviance = TRUE,
  include.nobs = TRUE, ...)
}
\arguments{
\item{model}{A model object.}

\item{include.aic}{Report Akaike's Information Criterion (AIC) in the GOF block?}

\item{include.rsquared}{Report R^2 in the GOF block?}

\item{include.maxrs}{Report maximal R^2 in the GOF block?}

\item{include.events}{Report the number of events in the GOF block?}

\item{include.nobs}{Report the number of observations in the GOF block?}

\item{include.missings}{Report number of missing data points in the GOF
block?}

\item{...}{Additional arguments for the \code{\link[base]{summary}} function.}

\item{include.zph}{Report proportional hazard test in the GOF block?}

\item{include.bic}{Report the Bayesian Information Criterion (BIC) in the GOF block?}

\item{include.loglik}{Report the log likelihood in the GOF block?}

\item{include.deviance}{Report the deviance in the GOF block?}
}
\value{
A \linkS4class{texreg} object.
}
\description{
\code{\link{extract}} methods for statistical models defined in the
\pkg{survival} package, in particular:
\itemize{
  \item \code{coxph} objects, which are created by the
    \code{\link[survival]{coxph}} function
  \item \code{coxph.penal} objects, which are created by the
    \code{\link[survival]{coxph}} function
  \item \code{clogit} objects, which are created by the
    \code{\link[survival]{clogit}} function
  \item \code{survreg} objects, which are created by the
    \code{\link[survival]{survreg}} function
  \item \code{survreg.penal} objects, which are created by the
    \code{\link[survival]{survreg}} function
}
}
\seealso{
\link{extract}

Other extract: \code{\link{extract,Arima-method}},
  \code{\link{extract,averaging-method}},
  \code{\link{extract,betareg-method}},
  \code{\link{extract,bife-method}},
  \code{\link{extract,biglm-method}},
  \code{\link{extract,brglm-method}},
  \code{\link{extract,brmsfit-method}},
  \code{\link{extract,censReg-method}},
  \code{\link{extract,coeftest-method}},
  \code{\link{extract,dynlm-method}},
  \code{\link{extract,ergm-method}},
  \code{\link{extract,ergmm-method}},
  \code{\link{extract,fGARCH-method}},
  \code{\link{extract,feglm-method}},
  \code{\link{extract,feis-method}},
  \code{\link{extract,felm-method}},
  \code{\link{extract,gamlss-method}},
  \code{\link{extract,glm-method}},
  \code{\link{extract,ivreg-method}},
  \code{\link{extract,lm-method}},
  \code{\link{extract,negbin-method}},
  \code{\link{extract,speedglm-method}},
  \code{\link{extract,speedlm-method}},
  \code{\link{extract.broom-method}},
  \code{\link{extract.btergm}},
  \code{\link{extract.forecast}},
  \code{\link{extract.lme4}}, \code{\link{extract.mfx}},
  \code{\link{extract.ordinal}}
}
\author{
Philip Leifeld
}
\concept{extract}
